Move back to Integers for all outcome enums
