+++ asset-manager-web/pom.xml
@@ -16,6 +16,17 @@
     <!--	1/2. https://docs.spring.io/spring-boot/docs/current/reference/htmlsingle/#build-tool-plugins-maven-packaging-->
     <packaging>jar</packaging>
 
+    <!-- jsk - want to force reuse of local snapshots. This is a fragment of the effective pom -->
+    <repositories>
+        <repository>
+            <snapshots>
+                <enabled>true</enabled>
+            </snapshots>
+            <id>central</id>
+            <name>Central Repository</name>
+            <url>https://repo.maven.apache.org/maven2</url>
+        </repository>
+    </repositories>
     <properties>
         <java.version>1.8</java.version>
     </properties>
@@ -114,12 +125,12 @@
         <dependency>
             <groupId>io.bdrc.am</groupId>
             <artifactId>audit-test-lib</artifactId>
-            <version>1.0-SNAPSHOT</version>
+            <version>0.9-SNAPSHOT-1</version>
         </dependency>
         <dependency>
             <groupId>io.bdrc.am</groupId>
             <artifactId>audit-test-interface</artifactId>
-            <version>1.0-SNAPSHOT</version>
+            <version>0.9-SNAPSHOT-1</version>
         </dependency>
     </dependencies>
