+++ audittool/.idea/modules.xml
@@ -2,6 +2,9 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
+      <module fileurl="file://$PROJECT_DIR$/audit-test-interface/audit-test-interface.iml" filepath="$PROJECT_DIR$/audit-test-interface/audit-test-interface.iml" />
+      <module fileurl="file://$PROJECT_DIR$/audit-test-lib/audit-test-lib.iml" filepath="$PROJECT_DIR$/audit-test-lib/audit-test-lib.iml" />
+      <module fileurl="file://$PROJECT_DIR$/audit-test-shell/audit-test-shell.iml" filepath="$PROJECT_DIR$/audit-test-shell/audit-test-shell.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/audittool.iml" filepath="$PROJECT_DIR$/.idea/audittool.iml" />
     </modules>
   </component>
