+++ audittool/audit-test-lib/src/main/java/io/bdrc/am/audit/audittests/FileSequence.java
@@ -1,6 +1,7 @@
 package io.bdrc.am.audit.audittests;
 
 import io.bdrc.am.audit.iaudit.ClassPropertyManager;
+import io.bdrc.am.audit.iaudit.LibOutcome;
 import io.bdrc.am.audit.iaudit.Outcome;
 import org.apache.commons.io.FilenameUtils;
 import org.slf4j.Logger;
