+++ audittool/audit-test-lib/src/main/java/io/bdrc/am/audit/audittests/ImageAttributeTests.java
@@ -1,6 +1,7 @@
 package io.bdrc.am.audit.audittests;
 
 import com.google.common.collect.Streams;
+import io.bdrc.am.audit.iaudit.LibOutcome;
 import io.bdrc.am.audit.iaudit.Outcome;
 import org.apache.commons.io.FilenameUtils;
 import org.slf4j.Logger;
