+++ audittool/audit-test-lib/src/main/java/io/bdrc/am/audit/audittests/NoFoldersInImageGroups.java
@@ -1,5 +1,6 @@
 package io.bdrc.am.audit.audittests;
 
+import io.bdrc.am.audit.iaudit.LibOutcome;
 import io.bdrc.am.audit.iaudit.Outcome;
 import org.slf4j.Logger;
